삭제 기능 구현

+++ src/main/java/com/ll/exam/article/repository/ArticleRepository.java
@@ -75,7 +75,7 @@ public class ArticleRepository {
         sql.update();
     }
 
-    public void delete(int id) {
+    public void delete(long id) {
         SecSql sql = myMap.genSecSql();
         sql
                 .append("DELETE FROM article")
